thesis/requirements.txt
attrs
Django==2.1.3
factory_boy
djangorestframework>=3.9.2
django-filter==2.1.0
psycopg2
pytest-django
factory_boy
freezegun
django-extensions==2.1.5
marshmallow
Faker
